build++
(Fix breakage from commit af8b8c2ce3f82eb8c08c6915eb09c88919bd31a5)

+++ src/core/capabilities/EditCapability.h
@@ -52,7 +52,7 @@ namespace Meta
 
             virtual void setDiscNumber( int newDiscNumber ) = 0;
 
-            virtual void setBpm( const float newBpm ) = 0;
+            virtual void setBpm( const qreal newBpm ) = 0;
 
             /** The track object should not store changed meta data immediately but cache the
             changes until endMetaDataUpdate() or abortMetaDataUpdate() is called */
